Move code to create validVc into a helper function.
